refactoring

+++ src/main/java/com/rapidsystems/soft/project/model/order/Order.java
@@ -33,7 +33,7 @@ public class Order {
 
     private Map<String, Integer> ordersPlates;
 
-    private Double price;
+    private int price;
 
 
 }
